add useClaimScreen hook
